backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
google-genai>=1.0.0
httpx==0.25.2
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

